First steps - not remotely complete, stalled out because of insufficient time to make all requisite changes before Marie leaves for Germany.
