+++ .gitignore
@@ -6,6 +6,7 @@
 *.blend1
 *.blend2
 *__pycache__*
+*.DS_Store
 
 ## -- Do not copy actual settings, only example for distribution -- ##
 Settings/*ettings*.json
