+++ Classes/Notes_TEMP.md
@@ -1,0 +1,3 @@
+Notes
+
+for saved name, file_name (NOT the same as Blender's filepath; it is not a complete path.)
