+++ blender_addon/bvp_tools.py
@@ -468,63 +468,13 @@ class DBSaveAction(bpy.types.Operator):
 		#	bvp.blend(script,pfile)
 		return {'FINISHED'}
 	def invoke(self,context,event):
-		global to_save
-		wm = context.window_manager
-		ob = context.object
-		act = ob.animation_data.action
-		## -- Compute parameters -- ##
-		## Frames
-		nframes = np.floor(act.frame_range[1])-np.ceil(act.frame_range[0])
-		## WordNet labels
-		wordnet_labels = [s.name for s in act.bvpAction.wordnet_label]
-		wordnet_frames = [s.frame for s in act.bvpAction.wordnet_label]
-		## Bounding box
-		if isinstance(ob.data,bpy.types.Armature):
-			# Get child objects, armatures have no position information
-			ob_list = [ob]+list(ob.children)
-		scn = context.scene
-		mn,mx = [],[]
-		for fr in range(int(np.floor(act.frame_range[0])),int(np.ceil(act.frame_range[1]))):
-		    scn.frame_set(fr)
-		    scn.update()
-		    mntmp,mxtmp = bvp.utils.blender.get_group_bounding_box(ob_list)
-		    mn.append(mntmp)
-		    mx.append(mxtmp)
-		min_xyz = np.min(np.vstack(mn),axis=0).tolist()
-		max_xyz = np.max(np.vstack(mx),axis=0).tolist()
-		#bvp.utils.blender.make_cube('bbox',min_xyz,max_xyz) # works. This shows the bounding box, if you want. 
-		bvp.utils.blender.grab_only(ob)
-		## Parent file
-		#pfile = act.bvpAction.parent_file
-		# The above value (pfile) is ignored for now. Need to eventually implement some way to take the contents 
-		# of the current file (group/action/whatever) and save them (append them) to another specfied file
-		# in the database. Currently NOT IMPLEMENTED.
-		thisfile = os.path.split(bpy.data.filepath)[1] #if len(bpy.data.filepath)>0 else pfile
-		if thisfile=="":
-			# Require saving in db-appropriate location 
-			raise NotImplementedError("Please save this file into %s before trying to save to database."%(os.path.join(dbpath,'Actions/')))
+		global bvpact
+		bvpact = Action.from_blender(context)
+		# (get docdict)
+		chk = dbi.query(**bvpact.docdict)
 
-		# Construct action struct to save
-		to_save = dict(
-			# Edited through UI 
-			act_name=act.name,
-			parent_file=thisfile,
-			wordnet_label=wordnet_labels,
-			wordnet_frames=wordnet_frames,
-			is_cyclic=act.bvpAction.is_cyclic,
-			is_translating=act.bvpAction.is_translating,
-			is_broken=act.bvpAction.is_broken,
-			is_armature=act.bvpAction.is_armature,
-			bg_interaction=act.bvpAction.bg_interaction,
-			obj_interaction=act.bvpAction.obj_interaction,
-			is_interactive=act.bvpAction.is_interactive,
-			is_animal=act.bvpAction.is_animal,
-			# Computed / assumed
-			nframes=nframes,
-			fps=act.bvpAction.fps,
-			min_xyz=min_xyz,
-			max_xyz=max_xyz,
-			)
+
+		## OLD
 		# Create database instance
 		dbi = bvp.bvpDB(port=dbport,dbname=wm.active_db)
 		# Check for existence of to_save in database
